+++ core/src/entry_point_callback.rs
@@ -2,6 +2,7 @@
 
 use casper_types::CLType;
 
+use crate::args::EntrypointArgument;
 use crate::call_def::CallDef;
 use crate::casper_types::bytesrepr::Bytes;
 use crate::{host::HostEnv, prelude::*, ContractEnv, OdraResult};
@@ -85,7 +86,10 @@ pub struct Argument {
 
 impl Argument {
     /// Creates a new instance of `Argument`.
-    pub fn new(name: String, ty: CLType) -> Self {
-        Self { name, ty }
+    pub fn new<T: EntrypointArgument>(name: String) -> Self {
+        Self {
+            name,
+            ty: T::cl_type()
+        }
     }
 }