+++ odra-macros/src/ast/wasm_parts_utils.rs
@@ -12,7 +12,11 @@ pub fn param_parameters(func: &FnIR) -> syn::Expr {
         .filter_map(Result::ok)
         .map(|(name, ty)| utils::expr::new_parameter(name, ty))
         .collect::<Vec<_>>();
-    parse_quote!(vec![#(#params),*])
+    if params.is_empty() {
+        parse_quote!(vec![])
+    } else {
+        parse_quote!(vec![#(#params),*].into_iter().filter_map(|x| x).collect())
+    }
 }
 
 pub fn param_access(func: &FnIR) -> syn::Expr {
@@ -32,9 +36,9 @@ pub fn param_ret_ty(func: &FnIR) -> syn::Expr {
 pub fn insert_arg_stmt(arg: &FnArgIR) -> syn::Stmt {
     let (name, ty) = arg.name_and_ty().unwrap();
     let args = utils::ident::named_args();
-
-    syn::parse_quote!(let _ = #args.insert(
+    syn::parse_quote!(odra::args::EntrypointArgument::insert_runtime_arg(
+        exec_env.get_named_arg::<#ty>(#name),
         #name,
-        odra::odra_casper_wasm_env::casper_contract::contract_api::runtime::get_named_arg::<#ty>(#name)
+        &mut #args
     );)
 }