+++ odra-macros/src/utils/ty.rs
@@ -89,10 +89,6 @@ pub fn entry_point_type() -> syn::Type {
     parse_quote!(odra::casper_types::EntryPointType)
 }
 
-pub fn parameter() -> syn::Type {
-    parse_quote!(odra::casper_types::Parameter)
-}
-
 pub fn group() -> syn::Type {
     parse_quote!(odra::casper_types::Group)
 }
@@ -106,6 +102,10 @@ pub fn schema() -> syn::Type {
 
 pub fn cl_typed() -> syn::Type {
     parse_quote!(odra::casper_types::CLTyped)
+}
+
+pub fn entry_point_arg() -> syn::Type {
+    parse_quote!(odra::args::EntrypointArgument)
 }
 
 pub fn cl_type() -> syn::Type {
@@ -162,10 +162,6 @@ pub fn entry_point_def_ty_constructor() -> syn::Type {
 
 pub fn entry_point_def_ty_public() -> syn::Type {
     parse_quote!(odra::contract_def::EntrypointType::Public)
-}
-
-pub fn entry_point_def_arg() -> syn::Type {
-    parse_quote!(odra::contract_def::Argument)
 }
 
 pub fn string() -> syn::Type {